+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+- Changed the migration from Neo 2.13.0 that adds the `groupId` column to the `neoblocktypes` table, to set each block type's `group` in the project config, instead of setting the `groupId` on the block type model and resaving the block type
+
 ## 2.13.2 - 2022-03-21
 
 ### Fixed